Create a library to check deployment
And also check initialization if ingress is not accessible, which will
skip the test.
